+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 <!-- markdownlint-disable MD024 -->
 # Changelog
 
+## `v0.9.0` (_aka_ 🐐)
+
+### Breaking changes
+
+* `class ABIContractExecutor` renamed to `ABIMethodCallStrategy` and moved from `graviton/blackbox.py` to `graviton/abi_strategy.py`
+
+### Added
+* `class Simulation` in `graviton/sim.y` unifies the ability to run an argument strategy and check that invariants hold using its `run_and_assert()` method
 
 ## `v0.8.0` (_aka_ 🦛)